余量计算接口 优化

+++ trfc-api/src/main/java/com/tianrui/api/resp/quality/sales/SalesBatchnumResp.java
@@ -112,11 +112,11 @@ public class SalesBatchnumResp extends BaseResp {
     /**
      * 过磅量
      */
-    private Long weighed;
+    private Double weighed;
     /**
      * 余量
      */
-    private Long margin;
+    private Double margin;
 	public String getId() {
 		return id;
 	}
@@ -261,18 +261,19 @@ public class SalesBatchnumResp extends BaseResp {
 	public void setMaterialname(String materialname) {
 		this.materialname = materialname;
 	}
-	public Long getWeighed() {
+	public Double getWeighed() {
 		return weighed;
 	}
-	public void setWeighed(Long weighed) {
+	public void setWeighed(Double weighed) {
 		this.weighed = weighed;
 	}
-	public Long getMargin() {
+	public Double getMargin() {
 		return margin;
 	}
-	public void setMargin(Long margin) {
+	public void setMargin(Double margin) {
 		this.margin = margin;
 	}
+
 
     
 }
